Fill in compat entries for every dependency
For some reason when I add ScatteredInterpolation.jl to my project it downgrades some other packages and insists on using an older version of Distances.jl. This doesn't happen when all of the dependencies are specified in the compat section of Project.toml (which is preferred practice anyways).

I've included Distances `v0.9` because it's only been `v0.10` since late 2020. Combinatorics has also been `v1` since 2019 and NearestNeighbors has been on `v0.4` for even longer.